refactoring: Extracts Cleanse module from Op class
